+++ buildpak.ps1
@@ -18,11 +18,11 @@ Remove-Item .\output\CHANGELOG.md
 7. Run the build using the pack task to create the NuGet package: `.\build.ps1 -tasks pack -CodeCoverageThreshold 0`.
 8. Upload the NuGet package to the PowerShell Gallery using the publish task: `.\build.ps1 -tasks publish -CodeCoverageThreshold 0`.
 #>
-$ver = "v0.2.1"
+$ver = "v0.3.0"
 git checkout main
 git pull origin main
 git tag -a $ver -m "Release version $ver Fix"
-"Fix: Issues #20 #21 #22 #23 #24"
+"Fix: Issues #25 #26"
 git push origin $ver
 # git tag -d $ver