+++ source/Public/Get-NetworkAudit.ps1
@@ -38,7 +38,7 @@ function Get-NetworkAudit {
             Position = 0
         )]
         [ValidateRange(1, 65535)]
-        [int[]]$Ports,
+        [Int32[]]$Ports,
         [Parameter(
             Mandatory = $true,
             ParameterSetName = 'Default',
@@ -63,11 +63,17 @@ function Get-NetworkAudit {
         # Begin Logging
         $Script:LogString += Write-AuditLog -Message "Begin Log"
         # Check if PSnmap module is installed, if not install it.
-        If (Get-Module -ListAvailable -Name "PSnmap") { Import-Module "PSnmap" } Else { Install-Module "PSnmap" -Force; Import-Module "PSnmap" }
-
+        # Tested Version:
+        # https://www.powershellgallery.com/packages/PSnmap/1.3.1 Updated: 7/18/2018
+        $params = @{
+            PublicModuleNames      = "PSnmap"
+            PublicRequiredVersions = "1.3.1"
+            Scope                  = "CurrentUser"
+        }
+        Initialize-ModuleEnv @params
         # Set default ports to scan
         if (!($ports)) {
-            [int[]]$ports = "21", "22", "23", "25", "53", "67", "68", "80", "443", `
+            [Int32[]]$ports = "21", "22", "23", "25", "53", "67", "68", "80", "443", `
                 "88", "464", "123", "135", "137", "138", "139", `
                 "445", "389", "636", "514", "587", "1701", `
                 "3268", "3269", "3389", "5985", "5986"
@@ -88,22 +94,28 @@ function Get-NetworkAudit {
                 # Validate the subnet format for IPv4 and IPv6
                 if (($subnet -match '^([0-9]{1,3}\.){3}[0-9]{1,3}(\/([0-9]|[1-2][0-9]|3[0-2]))?$') -or ($subnet -match '^s*((([0-9A-Fa-f]{1,4}:){7}([0-9A-Fa-f]{1,4}|:))|(([0-9A-Fa-f]{1,4}:){6}(:[0-9A-Fa-f]{1,4}|((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3})|:))|(([0-9A-Fa-f]{1,4}:){5}(((:[0-9A-Fa-f]{1,4}){1,2})|:((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3})|:))|(([0-9A-Fa-f]{1,4}:){4}(((:[0-9A-Fa-f]{1,4}){1,3})|((:[0-9A-Fa-f]{1,4})?:((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){3}(((:[0-9A-Fa-f]{1,4}){1,4})|((:[0-9A-Fa-f]{1,4}){0,2}:((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){2}(((:[0-9A-Fa-f]{1,4}){1,5})|((:[0-9A-Fa-f]{1,4}){0,3}:((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3}))|:))|(([0-9A-Fa-f]{1,4}:){1}(((:[0-9A-Fa-f]{1,4}){1,6})|((:[0-9A-Fa-f]{1,4}){0,4}:((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3}))|:))|(:(((:[0-9A-Fa-f]{1,4}){1,7})|((:[0-9A-Fa-f]{1,4}){0,5}:((25[0-5]|2[0-4]d|1dd|[1-9]?d)(.(25[0-5]|2[0-4]d|1dd|[1-9]?d)){3}))|:)))(%.+)?s*(\/([0-9]|[1-9][0-9]|1[0-1][0-9]|12[0-8]))?$')) {
                     # Create Network Scan Object
+                    $Script:LogString += Write-AuditLog -Message "Beggining scan of subnet $($subnet) for the following ports:"
+                    $Script:LogString += Write-AuditLog -Message "$(($ports | Out-String -Stream) -join ",")"
                     $NetworkAudit = Invoke-PSnmap -ComputerName $subnet -Port $ports -Dns -NoSummary -AddService
                     # Filter devices that don't ping as no results will be found.
                     $scan = $NetworkAudit | Where-Object { $_.Ping -eq $true }
                     # Write out information about the network scan.
-                    Write-Verbose "##########################################"
-                    Write-Verbose "Network scan for Subnet $Subnet completed."
-                    Write-Verbose "DHCP Server: $($DHCPServer)"
-                    Write-Verbose "Gateway: $($network.IPv4DefaultGateway.nexthop)"
-                    Write-Verbose "##########################################"
+                    $Script:LogString += Write-AuditLog -Message "##########################################"
+                    $Script:LogString += Write-AuditLog -Message "Network scan for Subnet $($Subnet) completed."
+                    $Script:LogString += Write-AuditLog -Message "DHCP Server: $($DHCPServer)"
+                    $Script:LogString += Write-AuditLog -Message "Gateway: $($network.IPv4DefaultGateway.nexthop)"
+                    $Script:LogString += Write-AuditLog -Message "##########################################"
+                    $Script:LogString += Write-AuditLog -Message "Creating $(($scan).count) output objects."
                     # For each device in the scan, get MAC ID vendor information and add it as a NoteProperty to the object.
                     $scan | ForEach-Object {
                         $org = ""
-                        $macid = ((arp -a $_.ComputerName | Select-String '([0-9a-f]{2}-){5}[0-9a-f]{2}').Matches.Value).Replace("-", ":") | Out-Null
-                        $macpop = $macid.replace(":", "") | Out-Null
-                        $macsubstr = $macpop.Substring(0, 6) | Out-Null
-                        $org = ($ouiobject | Where-Object { $_.assignment -eq $macsubstr })."Organization Name" | Out-Null
+                        $SaveErrorPref = $Script:ErrorActionPreference
+                        $Script:ErrorActionPreference = 'SilentlyContinue'
+                        $macid = ((arp -a "$($_.ComputerName)" | Select-String '([0-9a-f]{2}-){5}[0-9a-f]{2}').Matches.Value).Replace("-", ":")
+                        $macpop = $macid.replace(":", "")
+                        $macsubstr = $macpop.Substring(0, 6)
+                        $org = ($ouiobject | Where-Object { $_.assignment -eq $macsubstr })."Organization Name"
+                        $Script:ErrorActionPreference = $SaveErrorPref
                         Add-Member -InputObject $_ -MemberType NoteProperty -Name MacID -Value $macid
                         if ($org) {
                             Add-Member -InputObject $_ -MemberType NoteProperty -Name ManufacturerName -Value $org
@@ -112,11 +124,14 @@ function Get-NetworkAudit {
                             Add-Member -InputObject $_ -MemberType NoteProperty -Name ManufacturerName -Value "Not Found"
                         }
                     }
+                    $Script:LogString += Write-AuditLog -Message "Created $(($scan).count) output objects for the following hosts:"
+                    $Script:LogString += Write-AuditLog -Message "$(($scan | select-object "IP/DNS")."IP/DNS" -join ", ")"
+
                     # Normalize Subnet text for filename.
-                    $subnetText = $(($subnet.Replace("/", ".CIDR.")))
+                    $subnetText = $(($subnet.Replace("/", "_")))
                     # If report switch is true, export the scan to a CSV file with a timestamped filename.
                     if ($report) {
-                        $csv = "C:\temp\$((Get-Date).ToString('yyyy-MM-dd_hh.mm.ss')).$($env:USERDNSDOMAIN)_Subnet.$($subnetText)_DHCP.$($DHCPServer)_Gateway.$($network.IPv4DefaultGateway.nexthop).NetScan.csv"
+                        $csv = "C:\temp\$((Get-Date).ToString('yyyy.MM.dd_hhmm.ss')).$($env:USERDOMAIN)_Subnet.$($subnetText)_DHCP.$($DHCPServer)_GW.$($network.IPv4DefaultGateway.nexthop).NetScan.csv"
                         $zip = $csv -replace ".csv", ".zip"
                         $log = $csv -replace ".csv", ".AuditLog.csv"
                         return Build-ReportArchive -Export $scan -csv $csv -zip $zip -log $log -ErrorVariable BuildErr
@@ -133,9 +148,18 @@ function Get-NetworkAudit {
                 $csv = "C:\temp\$((Get-Date).ToString('yyyy-MM-dd_hh.mm.ss')).$($env:USERDNSDOMAIN)_HostScan.csv"
                 $zip = $csv -replace ".csv", ".zip"
                 $log = $csv -replace ".csv", ".AuditLog.csv"
-            return Build-ReportArchive -Export $results -csv $csv -zip $zip -log $log -ErrorVariable BuildErr
+                return Build-ReportArchive -Export $results -csv $csv -zip $zip -log $log -ErrorVariable BuildErr
             }
         }
+    <#
+        try {
+            ##
+        }
+        catch {
+            throw $_.Exception
+        }
+    #>
+
     } # Process Close
     end {
         return $results